Move get time function to helpFunction file
